ch19 optimizations

+++ src/main/java/com/quaxt/mcc/asm/ByteArray.java
@@ -1,6 +1,6 @@
 package com.quaxt.mcc.asm;
 
-public record ByteArray(int size, int alignment) implements TypeAsm {
+public record ByteArray(long size, long alignment) implements TypeAsm {
     public String suffix() {
         return "q";
     }
