+++ src/main/java/com/quaxt/mcc/asm/Data.java
@@ -1,7 +1,7 @@
 package com.quaxt.mcc.asm;
 
-public record Data(String identifier, int offset) implements Operand {
-    public Data(String identifier, int offset) {
+public record Data(String identifier, long offset) implements Operand {
+    public Data(String identifier, long offset) {
         this.identifier = identifier;
         this.offset = offset;
     }
