+++ src/main/java/com/quaxt/mcc/asm/PseudoMem.java
@@ -1,7 +1,7 @@
 package com.quaxt.mcc.asm;
 
-public record PseudoMem(String identifier, int offset) implements Operand {
-    public Operand plus(int offset) {
+public record PseudoMem(String identifier, long offset) implements Operand {
+    public Operand plus(long offset) {
         return new PseudoMem(identifier, this.offset + offset);
     }
 }
