+++ src/main/java/com/quaxt/mcc/asm/TypeAsm.java
@@ -4,9 +4,9 @@ import com.quaxt.mcc.Err;
 
 public sealed interface TypeAsm permits PrimitiveTypeAsm, ByteArray {
 
-    int size();
+    long size();
 
-    default int alignment() {
+    default long alignment() {
         return size();
     }
